Added docs, and changed version_t typedef to return a float.

+++ src/includes/API/IModule.h
@@ -157,7 +157,7 @@ namespace zAPI
     /// Defines a pointer to name() function, which return the module name as a string
     typedef std::string name_t();
     /// define a pointer to version() function, which return the api's version the module is maid for, as an integer.
-    typedef int         version_t();
+    typedef float       version_t();
 }
 
 #endif // IMODULE_H
